codeforce

+++ O.cpp
@@ -3,7 +3,7 @@ using namespace std;
 int main()
 {
     char x;
-    int a,b,result;
+long long int a,b,result;
     cin>>a>>x>>b;
     switch(x)
     {
